requirements.txt
py-dotenv
requests
os
json